Output structure automata descriptions.
When using an explicit structure-oriented automaton, we need to expose the
location of the automaton's "anchor" to the instrumenter so that we can
emit the automaton description there. This step was missed when we made
some changes to function instrumentation in the run-up to the paper, when
the structure automata were slightly neglected.