Set up locks framework
Moved my experimentation with building against TESLA into my fork's
source properly. This commit sets up a directory within the TESLA source
that I can run experiments etc. against and have it be built as part of
the TESLA build.

As well as this, this sets up a small example program that runs two
threads in mutual exclusion using a simple lock structure.

Additionally, do some environment setup so that docs don't break the
build on my machine for now, and add a convenience macro for TESLA field
access annotations.