Make TESLA build a function
This allows for the TESLA build stuff I've worked out to be pulled out
into its own file in the future so that multiple targets can be built
without duplication.

To do this I needed to do a fair bit of fiddling with the TESLA CMake
build I'd set up previously (largely to do with duplicated target and
dependency names). The current setup is such that source files can be
shared between several targets built by in the same directory (for
example, to share a common set of automata definitions between these
targets). The caveat here is that unfortunately the shared source files
are completely rebuilt for every target they're shared between.