Add some examples of misbehaving programs
Each of these will fail their TESLA assertions for not respecting the
lock semantics (in different ways). Also adds a small parameterisation
of the CMakeLists to allow for this.

Deduplicate examples of using lock automata

This splits up the harness and mocking from the behaviour of the program
with regard to the lock.