Add variant of CollectArgs
The existing code works on an automaton instance. The only thing
required from this automaton is its set of arguments. For the static
analysis we want to extract values from arguments, but do not always
have the relevant automaton. This allows us to pass the arguments in
manually to the analysis pass.