Add complete instrumentation removal
If all the usages in a manifest are deleted, then we should perform no
instrumentation at all (previously, lots of setup code was left in the
executables). This PR adds a new pass that just removes TESLA magic
function calls if safe to do so. Otherwise, the previous instrumentation
code will run unmodified.