Begin to structure instrumenter as an opt pass
Currently, the instrumenter is based on the structure of the `opt`
executable. To better integrate it with the new CADETS build
infrastructure, it should instead be in the form of a loadable LLVM
pass (as is the case with LOOM, for example).

This commit leaves the instrumenter in an intermediate stage with no
functionality - future commits will restore its code piece by piece.