Add instrumenter code back to the build process
The instrumenter code had previously been removed to test the process of
building a dynamically loadable `opt` pass. The code has been added back
to the build process in this commit, with the exception of the main
executable logic.

Doing this has required some changes to the way TESLA is built.
`libTeslaCommon` is now not statically linked against the LLVM
libraries (because the `opt` pass needs to be linked against the TESLA
library, and it can't also be statically linked against LLVM). This
means that any code linking against `libTESLACommon` must **first** link
against the necessary LLVM libraries.